pam: Use native PAM communication when possible
It's just better as it can be used both by terminal apps and other
interactive UIs, without any assumption on the terminal in use, this is
mostly due to the fact that a PAM client may be remote (as in SSH case)
and so we can't assume that is always using a terminal.

As per this, generalize the case trying to use the PAM APIs all the
times, while using the terminal specific features if that's possible.

So that ideally we can also share some of them with GDM ops.