pam: Add a simple go-loader pam module to wrap the go-generated modules
When a Go shared library is loaded from C, go starts various goroutine
(as init() at first) and if the loading code is then performing a fork
we end up having an undefined behavior and very likely, deadlocks.

To avoid this we need to ensure that the module is initialized after the
loading application has forked, and this for sure has happened when the
application calls the PAM functions that a module exposes.

As per this, here we re-implement the PAM modules functions and for
each action we load the module if it has not been loaded already,
otherwise we dload it and we redirect each call to the loaded library.

This is safe because a PAM application is always perform operations in
the same thread so once we've defined the "main" one we can create new
ones in the module safely.

By using this wrapper the pam_authd.so module can now loaded using:

 auth required pam_go_loader.so pam_authd.so socket=/run/some.socket

Added a pam-auth-update generator to make this easier.