model: Simplify stage change logic
Avoid repeating the same operations multiple times or handling change
stage messages when not needed.

To make this to work we need to ensure that the initial state is not
marked as user selection. We do not define a new stage value since it
would be unused, and that's good that is not.