Add Debian packaging
Build the authd package in a single binary package containing the
daemon, PAM & NSS modules.

Add hooks to ensure PAM & NSS are properly configured and un-configured.

Most of this was inspired by adsys and aad-auth packaging where we also
ship PAM (and NSS) modules.

d/copyright is still incomplete but it's an arduous task we can handle
afterwards.

Fixes UDENG-1781