Refactor daemon stop logic in testutils
We used to rely on two go routines, one for running the daemon and one
for stopping it. This created a race because cmd.Run() does not work
with cmd.Process.Signal() due to read/write blocks. Instead, now we use
CommandContext and override the cmd.Cancel function to send a SIGTERM
rather than a SIGKILL when the context is cancelled.