pam: Add proto file to define stage enum type and use it
In some implementations (e.g. gdm) we need to communicate with the UI
the stage that is expected to be shown. Since we have the same already
defined in the PAM module for local implementation, we can just expose
the same values in a newly defined pam protocol so that can be reused by
gdm too.

As per this, add a Stage enum type to the protocol and reuse these
values inside the pam module.

We just need to also define an undefined state to have an invalid value.