authd: Define AuthenticationData in proto and use to parse and encode values
The IsAuthenticated() rpc method used to send back to the broker the
challenge response from the UI, this is something that can be extended
depending on authentication methods, but right now we were defining
the return values manually.

This is not optimal, and right now implies hand-writing JSON from PAM
side without a proper data check.

To prevent this:
 - Add IARequest AuthenticationData type to the proto defined with the
   possible values right now, all optional.
 - Use the AuthenticationData in PAM side to send the IsAuthenticated()
   request to the daemon.
 - Use it in authd's brokers level to encode the result as the JSON that
   will be sent to the actual broker.

Note that we still continue passing the JSON as it is to the brokers.

However, this required fixing the examplebroker's parsing of requested
string, since according to the specs, the authenticationData should be
a JSON value, but before we were not properly formatting the boolean
items, and so sending to brokers invalid JSON.