pam/internal/gdm: Add main test function to double check we don't leak
We do leak checks when every test is over, however due to the fact that
some data will be finalized only after GC has fully ran, we may still
show some false-positive leaks.

As per this, double check this is not the case by running the leak
sanitizer checks again just before exiting. This should be the normal
behavior in any program usin ASAN, but it's not the case in Go, sadly