pam/gdmmodel_test: Ensure we only send auth data events when authentication has started
Do not assume that being in challenge stage means we are ready for
authentication as this may not be true, but instead wait for
authentication start event.

We may also wait for both things to have happened, but per se this is
not required as authentication may also happen because of broker
dictates that at any time.