pam: Only try to run the gdm mode if module is started by gdm-authd service
While checking for the extension support is a requirement, this is also
something that basic upstream gdm now supports by default, so let's
check that the authd is launched.

This also prevents to start the GDM mode when we're actually using the
normal password authentication via gdm and authd is configured to run as
default access mode.