pam: Add C implementation for an exec-module with private dbus support
We know that go isn't the best as being a shared library that is loaded
by C code, and it breaks the PAM stack quite often.

In order to perform more complex PAM communications by still using the
go code we have done already, we can instead execute the go code as an
exec client and then use DBus to communicate with it.

To do this, add a quite simple glib-based module implementation that:
  - Launches a binary passing to it the address of a private DBus
    server that the module itself starts.
  - Allows connections on such private server coming only from the
    launched child.
  - Exposes methods to perform the pam operations we care about, acting
    as a proxy between PAM and the client using DBus as transport.

In order to test this, add a go client implementation that can be
launched by the module and that can perform dbus operations on the
module by using the service file arguments.

So, test all this using an actual PAM transaction that:
  - Creates service files using our commands syntax as arguments
  - Loads the C module
  - Verifies that all the operations are doing the actions we expect
    and the values that we set from Go-client side are visible from
    the PAM-application side.