pam/go-exec: Add support for string conversations
The module is now feature complete to handle most of cases we care about
allowing to talk with the outer world.

As per this, add integration tests too ensuring this works as expected.
It requires few adjustments to the test client, to allow calling
variadic functions, parse more complex return values and and properly
compare errors.