pam/go-exec: Be consistent with libpam, potentially returning NULL data
If some module data is unset or NULL, libpam should in theory return
Pam.ErrNoModuleData, but sadly this not true if some data has been
removed after being set. See [1].

So, be consistent with what libpam does also in our exec module, and
just return what libpam does, so that we will have the same behavior
both when loading the library directly (as gdm does) or when PAM is
used through the exec module.

[1] https://github.com/linux-pam/linux-pam/pull/780