pam/go-exec: Ensure that GLib unix signal handlers are called in the right stack
When loading this module as part of a go program, we should ensure that
the signal handlers are called with the proper signal stack or we'd
end up accessing Go memory that should not be touched by the C code.

As per this, force using SA_ONSTACK for the SIGCHILD signal handler.

This could be done in all the cases, but let's limit to the go-tests
case for now, as there's no much point of doing this all the times from
our side.