Check AUTHTOK before prompting user for password
We used to always ask the user for the new password but, if the system
had pwquality enabled, the user would need to input the new password 4
times in a row (and the second pair didn't need to match the first).

Now, if AUTHTOK was already set, we can use it instead (meaning that
pwquality has already evaluated the password correctness and allowed it
to proceed).