Fix PAM chauthtok behavior
By having our module configured as `sufficient` and returning
pam.Success on the preliminary checks for pam_sm_chauthtok, we were
ignoring pam_unix and affecting the correct behavior of the pwquality
module, since the user was never asked for the current password (this
was affecting local users mostly).

Now, we check if the user is local (by looking at the default broker)
and return pam.Ignore to proceed with the stack (or Success if the user
is a remote one, since our module requires authentication in order to
change the password).