pam/integration-tests: Add test checking for mistyping password on reset
This codepath wasn't tested in the interactive client, but the native
PAM code has some special codepath to handle this, so it requires
testing. Thus prepare the test for cli first.

Again, we need to add an extra needs-reset user to be able to perform
this without clashing with other tests.