pam/gdmmodel: Support cancelling authentication all the times
Cancellation needs to be possible everytime we're in challenge mode, but
this is handled already by the authentication backend so don't bother
adding the logic here.

Since in this way we'll prevent being able to cancel authentications
that are in "wait" state and so when we're not anymore waiting for user
authentication request, but only for the broker reply.