pam/authentication: Start new auth only after cancellation is notified
In case we're still in the process of cancelling the authentication we
should not continue with the current action until we're notified that
the cancellation is completed.

This is needed because otherwise we may end up into this situation:
 - reselectAuthMode
 - AuthModeSelected
 - ...
 - startAuthentication
 - isAuthenticatedResultReceived: cancelled

Or a bit less extreme, but still getting the cancelled event while we're
selecting the authentication mode. And this mis-order causes troubles
especially when passing the data to gdm.

To avoid this, add a channel to wait for the authentication cancellation
to be completed and only once we've received a response we can proceed
with next actions.

Due to the blocking nature of channels we need to use more async
functions now and add a few synchronization points

Co-Authored-By: Denison Barbosa <denison.barbosa@canonical.com>