brokers/broker: Wait for the IsAuthenticated to be delivered before cancelling it
When quickly calling IsAuthenticated followed by context cancellation,
we may end up in a situation in which we cancel the request before we've
actually placed one on the broker.

This is due to the fact that the dbus request may take some time to be
delivered and being stored by a remote broker, while the cancellation
may happen very quickly.

So, to prevent this, in case the context has been cancelled, let's wait
"a bit" before sending the cancellation request, so that we can be sure
that the `IsAuthenticate` request has been delivered before.

This is far from being nice, but it's the best way to achieve this
result given the API constraints that we have, and the fact that go-dbus
can't advertise us if a message has been delivered, without waiting for
the whole result.

See example log of this failure at: https://www.pastery.net/afkmuy/