brokers/broker: Wait to do IsAuthentication requests for the same session
The IsAuthentication request on client side may happen even while we're
still cancelling the previous one, this is because the client receives
the cancellation error as just after the calling context is cancelled,
and not when the server side has actually completed and returned the
request. This is due to the nature of the grpc.ClientConnInterface
used, and there's no simple way to control this client-side.

To workaround this, on server side, do not permit that multiple auth
requests are performed for the same session, using a cond to wait for
a potentially previous one to be completed and return only once this is
possible.

This is far from optimal, but we have no other way at client side to
know that the server has finished with the all the previous requests.