pam/userselection: Do not allow changing user name if provided by PAM
In case the user name is already provided by PAM, do not permit
changing the user or some tools (su...) may accept silently the
user change and let any user to log in with the credential of another
user. The user view is now only presented in case no PAM user is
provided, but by default it's not enabled or reactive to events.

Fixes: GHSA-x5q3-c8rm-w787