pam/integration-tests: Programmatically generate the vhs tape headers
We repeat the same for all the files and it's hard to define specific
values, so avoid repeating the same for each file but instead generate
the contents programmatically so that it can be changed per test.

This generalization makes adding new tests easier and will permit to
change the tape data before running it.