pam/adapter: Only draw a model when it's also focused
In the tests that we have we now rely on the fact that when something is
on the screen, then it's also possible to type on it, while this is not
really the case in our bubbletea implementation, because we may end up
drawing and only then we focus a view.

Prevent this, by only showing a model when it's also focused (or ready
to draw), so that there's no risk of showing something and not being
able to get its input.