users/db/migration: Save a database of the group file
And try to restore it on error.

Mimic more what gpasswd does, so save a backup file and try to restore
it in case of failure.

We don't backup the backup file itself in case, we just try to keep a
copy of it to restore if backing up fails