GitHub issue template: Fix sudo password prompt not visible
When using sudo in the subshells of the command substitutions in the
heredoc, the password prompt is not visible, so the command seems to
hang, while actually it's waiting for the user to input their password.

To work around that, use `sudo -v` before the heredoc, so that the
user's credentials are cached and sudo doesn't ask for them again
(assuming that credential caching hasn't been disabled).