nativemodel: Be coherent with default logic when no broker is available
In case no broker is available the module returns a PAM_AUTHINFO_UNAVAIL
error, and os we should do here if polkit is used and a local user is
selected.

As per our default PAM configuration, this will lead to ignoring the
error, just as in the current case, but gives more control on the pam
module configuration, if someone wants to.

Partially reverts 48bad6a724ca91d149824b9061d38e7bef5c1b01