users/localentries/preauth: Do not deadlock on concurrent reads/writes
While a RWMutex may allow read while read-locked, we may still end up
in a situation in which we have a write lock and we are trying to read
a value. And this can be easily tested by an upcoming change.

To prevent this starvation to happen, let's just read-lock once when
getting an user from its UID