[WIP][MOVE UP] users/manager: Try more to find unique UIDs/GIDs
In case we have concurrent requests, as in the initial test, we may end
up failing updating the user, as we may have generate an UID/GID that is
being registered in the DB.

So in case this fails, let's just retry again, making our machinery to
look again for a new GID/UID if possible or to reuse the one we have
already set.

We still have issues in running concurrent updates for the same user
though, but this is due to the database, and so maybe to be handled in
a different place:

  2025/06/26 03:54:31 DEBUG Updating local groups for user "authd-test-maybe-pre-check-user93", new groups: [authd-test-group93 authd-test-other-group93], old groups: []
  2025/06/26 03:54:31 DEBUG Adding "authd-test-maybe-pre-check-user93" to local groups: [authd-test-group93 authd-test-other-group93]
  2025/06/26 03:54:31 DEBUG Removing "authd-test-maybe-pre-check-user93" from local groups: []
  2025/06/26 03:54:31 DEBUG Updating entry of user "authd-test-maybe-pre-check-user93" (UID: 70)
  2025/06/26 03:54:31 DEBUG Nothing to do, groups are equal
  2025/06/26 03:54:31 DEBUG Forgetting tracked user name "authd-test-maybe-pre-check-user93" for UID 70: true
  2025/06/26 03:54:31 DEBUG Updating user authd-test-maybe-pre-check-user93
  2025/06/26 03:54:31 DEBUG Updating entry of group "authd-test-maybe-pre-check-user93" ({Name:authd-test-maybe-pre-check-user93 GID:70 UGID:authd-test-maybe-pre-check-user93})
  2025/06/26 03:54:31 DEBUG Adding user 70 to group 70

  failed to update user "authd-test-maybe-pre-check-user93":
  failed to add user to group: FOREIGN KEY constraint failed
  (user with UID 70 does not exist) (group with GID 70 does not exist)

One option is to retry also in case we got such error, although I feel
that the database should not fail here, since we - really - in theory
just have saved it.

Maybe that's because we're not locking it? Indeed we should use the db
mutex for various reasons, but it doesn't seem to help either.