users/idgenerator: Simplify the used IDs cleanup logic
We used to drop the IDs in a single shot but this implied adding a
further method to the interface that had had no clear scope and was too
dangerous to use, so use a cleanup function instead so that we can
achieve the same, but schedule to drop each element once we're done with
it.

Also, given we use this struct in a locked environment we can avoid any
further locking, so dropping the mutex. The race tests would tell us if
in future this will be needed